# serve/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Blog App",
        "description": "Blog with back-end laravel and front-end flutter",
        "version": "1.0"
    },
    "paths": {
        "/api/register/": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register",
                "operationId": "73cf8b99c1004324ef4988fa9ea65cdf",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "password"
                                            },
                                            "password_confirmation": {
                                                "type": "password"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Admin",
                                    "email": "admin@admin.com",
                                    "password": "123456",
                                    "password_confirmation": "123456"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Admin"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "admin@admin.com"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2022-12-11T09:25:53.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2022-12-11T09:25:53.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "2|RlmF3V1BwHtXErI9md4OBu9ZvyjNmtCjt1Mnfv4f"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The email has already been taken."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login/": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login",
                "operationId": "5eff322f7def3bb9b9e045d9398035ab",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "password"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "admin@admin.com",
                                    "password": "123456"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Admin"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "admin@admin.com"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "null"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2022-12-11T09:25:53.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2022-12-11T09:25:53.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "2|RlmF3V1BwHtXErI9md4OBu9ZvyjNmtCjt1Mnfv4f"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout/": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Logout",
                "operationId": "0d1472f2c3c9109e6cbf60280872337a",
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logout success."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/user/": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get Detail User",
                "operationId": "83a260237694439c3ac3efa1c601de60",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Admin"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "admin@admin.com"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "null"
                                                },
                                                "email_verified_at": {
                                                    "type": "string",
                                                    "example": "null"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2022-12-11T09:25:53.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2022-12-11T09:25:53.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update",
                "operationId": "4b667476c1b93fd8888ccf72cc5f5664",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "image",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "file"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "image": {
                                                "type": "file"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Admin",
                                    "image": "null"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User updated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/post/{id}": {
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Delete Post",
                "operationId": "6d25b4d3f1749560f4d0d0ae67e41858",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Post deleted."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}